+++ include/MyECS/CmptLocator.h
@@ -35,7 +35,7 @@ class CmptLocator {
   bool HasWriteCmptType() const noexcept;
 
  private:
-  size_t GenHashCode() const noexcept;
+  void UpdateHashCode() noexcept;
 
   CmptAccessTypeSet cmptTypes;