+++ include/MyECS/detail/Archetype.h
@@ -14,10 +14,7 @@
 
 #include <MyContainer/Pool.h>
 
-#include <MyTemplate/TypeID.h>
 #include <MyTemplate/Typelist.h>
-
-#include <map>
 
 namespace My::MyECS {
 class EntityMngr;
@@ -41,17 +38,16 @@ class Archetype {
 
   // auto add Entity
   static Archetype* New(RTDCmptTraits&, Pool<Chunk>* chunkPool,
-                        const CmptType* types, size_t num);
+                        Span<const CmptType> types);
 
   // auto add Entity
   template <typename... Cmpts>
   static Archetype* Add(const Archetype* from);
   static Archetype* Add(RTDCmptTraits&, const Archetype* from,
-                        const CmptType* types, size_t num);
+                        Span<const CmptType> types);
 
   // auto add Entity
-  static Archetype* Remove(const Archetype* from, const CmptType* types,
-                           size_t num);
+  static Archetype* Remove(const Archetype* from, Span<const CmptType> types);
 
   // Entity + Components
   std::tuple<std::vector<Entity*>, std::vector<std::vector<CmptAccessPtr>>,
@@ -110,7 +106,7 @@ class Archetype {
   size_t ChunkCapacity() const noexcept { return chunkCapacity; }
 
   // add Entity
-  static CmptTypeSet GenCmptTypeSet(const CmptType* types, size_t num);
+  static CmptTypeSet GenCmptTypeSet(Span<const CmptType> types);
   template <typename... Cmpts>
   static CmptTypeSet GenCmptTypeSet();
 
@@ -121,7 +117,7 @@ class Archetype {
 
   size_t Offsetof(CmptType type) const { return type2offset.at(type); }
 
-  static bool NotContainEntity(const CmptType* types, size_t num) noexcept;
+  static bool NotContainEntity(Span<const CmptType> types) noexcept;
 
   friend class EntityMngr;